tidy up examples, fix current_playing
