+++ Cargo.toml
@@ -85,29 +85,24 @@ __sync = []
 features = ["cli"]
 
 [[example]]
-name = "album"
+name = "client_credentials"
 required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/album.rs"
+path = "examples/client_credentials.rs"
 
 [[example]]
-name = "current_user_recently_played"
+name = "code_auth"
 required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/current_user_recently_played.rs"
+path = "examples/code_auth.rs"
+
+[[example]]
+name = "code_auth_pkce"
+required-features = ["env-file", "cli", "client-reqwest"]
+path = "examples/code_auth_pkce.rs"
 
 [[example]]
 name = "oauth_tokens"
 required-features = ["env-file", "cli", "client-reqwest"]
 path = "examples/oauth_tokens.rs"
-
-[[example]]
-name = "track"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/track.rs"
-
-[[example]]
-name = "tracks"
-required-features = ["env-file", "cli", "client-reqwest"]
-path = "examples/tracks.rs"
 
 [[example]]
 name = "with_refresh_token"
