--- examples/album.rs
@@ -1,38 +1,0 @@
-use rspotify::{model::Id, prelude::*, ClientCredentialsSpotify, Credentials};
-
-#[tokio::main]
-async fn main() {
-    // You can use any logger for debugging.
-    env_logger::init();
-
-    // Set RSPOTIFY_CLIENT_ID, RSPOTIFY_CLIENT_SECRET and
-    // RSPOTIFY_REDIRECT_URI in an .env file or export them manually:
-    //
-    // export RSPOTIFY_CLIENT_ID="your client_id"
-    // export RSPOTIFY_CLIENT_SECRET="secret"
-    //
-    // These will then be read with `from_env`.
-    //
-    // Otherwise, set client_id and client_secret explictly:
-    //
-    // ```
-    // let creds = Credentials {
-    //     client_id: "this-is-my-client-id".to_string(),
-    //     client_secret: "this-is-my-client-secret".to_string()
-    // };
-    // ```
-    let creds = Credentials::from_env().unwrap();
-
-    let mut spotify = ClientCredentialsSpotify::new(creds);
-
-    // Obtaining the access token. Requires to be mutable because the internal
-    // token will be modified. We don't need OAuth for this specific endpoint,
-    // so `...` is used instead of `prompt_for_user_token`.
-    spotify.request_token().await.unwrap();
-
-    // Running the requests
-    let birdy_uri = Id::from_uri("spotify:album:0sNOF9WDwhWunNAHPD3Baj").unwrap();
-    let albums = spotify.album(birdy_uri).await;
-
-    println!("Response: {:#?}", albums);
-}
